[AA] Add symbol tests for function inside dangling annotation
The bug in the PSI-based version is that
the symbol locations should be LOCAL instead of TOP_LEVEL / CLASS.

Also, the `firSymbol`-based version returns different results:
the location is LOCAL (which is correct), but the `callableId` is
not null (should be null for local declarations).

Both of these problems will be fixed in the following commit.

KT-75534