[AA] Restrict project structure providers and symbol providers to module content
For now, I'm going with the `KaNotUnderContentRootModule` for
`.kotlin_metadata` and other platform-excluded files. That's in contrast
to the initially agreed-upon solution of providing a dangling file
module for the excluded file. This has the following reasons:

1. Simplicity: It's easy to implement the restriction filtering as a
   simple `virtualFile in module.contentScope` check. If we want to
   provide dangling file modules for library-excluded files
   specifically, we cannot just check the content scope, but need to
   make sure that platform exclusion was actually the reason for the
   failed content scope check. This makes the implementation more
   convoluted.
2. Correct modeling: If we truly cannot find a `KaModule` which has the
   file in its content scope, for all intents and purposes from the
   perspective of the Analysis API, the file is NOT part of the project
   content. As such, `KaNotUnderContentRootModule` fits best.
   - This decision already affects the code. For example, in
     `SymbolKotlinAsJavaSupport.findClassOrObjectDeclarations`,
     candidates are filtered by whether they're from a source or binary
     library. We need to exclude `.kotlin_metadata` files here if
     they're not in the platform content, since symbol light classes are
     the Java view on *Kotlin*. But if we provide a dangling file module
     for the `.kotlin_metadata` file with the original library as a
     context module, the `isFromSourceOrLibraryBinary` check will see
     the `.kotlin_metadata` file as a binary library file. Then the
     declaration will be included in the result.

It is true that when we return a `KaNotUnderContentRootModule` for the
`.kotlin_metadata` file, the IntelliJ highlighting in the file doesn't
work anymore. But the file needs to be opened directly to see this.
Given that we're dealing with a broken project structure (where the
`.kotlin_metadata` file shouldn't even exist), I can't see this
legitimately bothering our users.

Computing the module's content scope during `getModule` is slightly
risky, as we might run into a recursive loop when content scope refiners
themselves call `getModule`. But the work in this commit recognizes that
and adds specific restrictions to content scope refiners and resolve
extension providers.

^KT-74907 fixed