[AA] Additional tests for dangling annotations
Cover the case of classes and properties.

Note: a class or property inside a dangling annotation directly
is not parsed as a declaration, just as a bunch of error elements
(both on file and member level). So in these tests we cheat
and put the class/property inside a lambda first.

KT-75534