[IR] Reuse the signature when copying an inline function
The signature of a copied inline function is not going to be used
anywhere. During the inlining process we are interested only in
the body of the inline function. Also, during the serialization
of a preprocessed inline function, the signature can repeat because
this copy will be placed in a different list than everything else.

#KT-76425