[PL] Do not report an error if `reflectionTargetSymbol` is missing
With new rich references we have an additional logic in PL engine.
In this commit this logic was refined.
1. Do not create stub declarations for reflection property if
these declarations are already bound. Otherwise, we will end
up with a "missing declaration" error, although all declarations are
in place.
2. Do not log by default an error when constructing classes
for a reference with PL error. An error will be reported automatically
during the runtime.

#KT-76712 Fixed
#KT-76711 Fixed