[Analysis API] Fix compilation of common sources without JVM counterpart
While 'KaFirCompilerFacility' supports common code compilation after
KT-72476, previously it did so by compiling common code as it was a part
of the JVM implementation module. That worked, but sometimes, there is
no JVM module at all – only a common part that is directly compiled for
multiple platforms.

This commit fixes compilation of common-only modules by creating a
custom dangling file module. Effectively, the created module emulates
the missing JVM counterpart. Source dependencies for it are produced by
collecting all JVM counterparts of the module's dependencies. Library
dependencies, on the other hand, are inferred from modules that depend
on the common module.

For the fix, extra dependency support was added for dangling file
modules.

^KT-76457 Fixed