[K/N] Introduce ClangArgs.clangArgsForKonan(C|Cxx)Sources
There is `ClangArgs.clangArgsForKonanSources`, but it is specific to
C++ (clang++) and fails to mention this in the name.

This commit introduces two other properties to support the C case and
make the distinction clear.

It also makes public the language-agnostic part,
`ClangArgs.clangArgsSpecificForKonanSources`, just to make things more
flexible.