[K/N][build] Make NativeInteropPlugin ignore some warnings in C code
This code is mostly generated by cinterop, and it is known to make the
C compiler emit some warnings.

This commit ignores a few categories of such warnings to make the build
log cleaner.

Another ignored warning is about unused compiler arguments on Windows
(which is caused by imprecise implementation of ClangArgs.Jni -- it puts
linker args and compiler args into the same list).

^KT-77091